Project/stable-requirements.txt
certifi==2019.9.11
cycler==0.10.0
decorator==4.4.1
EasyProcess==0.2.7
kiwisolver==1.1.0
matplotlib==3.1.1
mss==4.0.3
networkx==2.4
numpy==1.17.4
opencv-python==4.1.1.26
Pillow==6.2.1
pyparsing==2.4.5
pytesseract==0.3.0
python-dateutil==2.8.1
six==1.13.0
StringDist==1.0.9
wincertstore==0.2
